requirements.txt
# Core dependencies
streamlit==1.32.0
python-dotenv==1.0.0

# Database and vector store
numpy==1.26.4
chromadb==0.4.22
duckdb==0.9.2

# LangChain related
langchain==0.1.0
langchain-core==0.1.22
langchain-community==0.0.20
langchain-google-genai==2.0.11  # Updated version
google-generativeai==0.3.2

# Web scraping
selenium==4.18.1
beautifulsoup4==4.12.0
webdriver-manager==4.0.1
lxml==5.1.0

# Utilities and parsing
watchdog==3.0.0
certifi==2024.2.2
urllib3==2.2.0#!/bin/bash

# Python 버전 확인
python_version=$(python -V 2>&1 | grep -Po '(?<=Python )(.+)')
echo "Installing packages for Python ${python_version}"

# 가상환경 재생성
rm -rf venv
python -m venv venv
source venv/bin/activate

# pip 업그레이드
pip install --upgrade pip setuptools wheel --trusted-host pypi.org --trusted-host files.pythonhosted.org

# SSL 관련 패키지 설치
pip install certifi --trusted-host pypi.org --trusted-host files.pythonhosted.org
pip install 'requests[security]' --trusted-host pypi.org --trusted-host files.pythonhosted.org

# 기본 의존성 개별 설치
pip install numpy==1.26.4 --trusted-host pypi.org --trusted-host files.pythonhosted.org
pip install chromadb==0.4.22 --trusted-host pypi.org --trusted-host files.pythonhosted.org

# LangChain 관련 패키지 개별 설치
pip install langchain==0.1.0 --trusted-host pypi.org --trusted-host files.pythonhosted.org
pip install langchain-core==0.1.22 --trusted-host pypi.org --trusted-host files.pythonhosted.org
pip install langchain-community==0.0.20 --trusted-host pypi.org --trusted-host files.pythonhosted.org
pip install langchain-google-genai==2.0.11 --trusted-host pypi.org --trusted-host files.pythonhosted.org
pip install google-generativeai==0.3.2 --trusted-host pypi.org --trusted-host files.pythonhosted.org

# 나머지 requirements.txt 설치
pip install -r requirements.txt --no-deps --trusted-host pypi.org --trusted-host files.pythonhosted.org
requests[security]==2.31.0

# SSL and security
pyOpenSSL==24.0.0
cryptography==42.0.0

# Vector operations
scikit-learn==1.4.1.post1